# api/openapi/zke-server.v1.yaml
openapi: 3.1.0
info:
  title: ZKE Server API
  version: v1
  description: |
    ZKE 用户管理、资源管理、Cluster 接入管理和内部 Agent 注册使用的 HTTP API 契约。
    所有 JSON 时间均为 RFC 3339，并由当前 API 以 UTC+8 偏移返回。
servers:
  - url: /
    description: Console 与 API 使用同源入口；本地开发通过前端代理转发
...
paths:
...
  /api/v1/tenants:
    get:
      tags: [Resources]
      operationId: listTenants
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/ListLimit"
        - $ref: "#/components/parameters/ListOffset"
        - $ref: "#/components/parameters/ListSearch"
        - name: status
          in: query
          schema:
            type: string
            enum: [active, suspended]
      responses:
        "200":
          description: 可见 Tenant
          content:
            application/json:
              schema:
                type: object
                required: [tenants, pagination]
                properties:
                  tenants:
                    type: array
                    items:
                      $ref: "#/components/schemas/Tenant"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
    post:
      tags: [Resources]
      operationId: createTenant
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        $ref: "#/components/requestBodies/CreateNamedResource"
      responses:
        "201":
          description: Tenant 已创建
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        "200":
          description: 幂等恢复
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        "409":
          $ref: "#/components/responses/Conflict"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/tenants/{tenant_id}:
    parameters:
      - $ref: "#/components/parameters/TenantID"
    get:
      tags: [Resources]
      operationId: getTenant
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Tenant 详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthenticated"
    put:
      tags: [Resources]
      operationId: updateTenant
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateScopedResourceRequest"
      responses:
        "200":
          description: Tenant 已更新
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        "409":
          $ref: "#/components/responses/Conflict"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
    delete:
      tags: [Resources]
      operationId: deleteTenant
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Confirmation"
      responses:
        "200":
          description: Tenant 已停用，下属接入身份已撤销
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/tenants/{tenant_id}/projects:
    parameters:
      - $ref: "#/components/parameters/TenantID"
    get:
      tags: [Resources]
      operationId: listProjects
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/ListLimit"
        - $ref: "#/components/parameters/ListOffset"
        - $ref: "#/components/parameters/ListSearch"
        - name: status
          in: query
          schema:
            type: string
            enum: [active, suspended]
      responses:
        "200":
          description: 可见 Project
          content:
            application/json:
              schema:
                type: object
                required: [projects, pagination]
                properties:
                  projects:
                    type: array
                    items:
                      $ref: "#/components/schemas/Project"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
    post:
      tags: [Resources]
      operationId: createProject
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        $ref: "#/components/requestBodies/CreateNamedResource"
      responses:
        "201":
          description: Project 已创建
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "200":
          description: 幂等恢复
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "409":
          $ref: "#/components/responses/Conflict"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/projects/{project_id}:
    parameters:
      - $ref: "#/components/parameters/ProjectID"
    get:
      tags: [Resources]
      operationId: getProject
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Project 详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
    put:
      tags: [Resources]
      operationId: updateProject
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateScopedResourceRequest"
      responses:
        "200":
          description: Project 已更新
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "409":
          $ref: "#/components/responses/Conflict"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
    delete:
      tags: [Resources]
      operationId: deleteProject
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Confirmation"
      responses:
        "200":
          description: Project 已停用，下属接入身份已撤销
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/projects/{project_id}/clusters:
    get:
      tags: [Resources]
      operationId: listClusters
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/ProjectID"
        - $ref: "#/components/parameters/ListLimit"
        - $ref: "#/components/parameters/ListOffset"
        - $ref: "#/components/parameters/ListSearch"
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, active, revoked]
      responses:
        "200":
          description: Project Cluster 列表
          content:
            application/json:
              schema:
                type: object
                required: [clusters, pagination]
                properties:
                  clusters:
                    type: array
                    items:
                      $ref: "#/components/schemas/ClusterAggregate"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: zke_session
    enrollmentToken:
      type: http
      scheme: bearer
      bearerFormat: Agent enrollment token
    installationToken:
      type: http
      scheme: bearer
      bearerFormat: Agent installation token
  parameters:
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 128
...
    TenantID:
      name: tenant_id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/UUID"
    ProjectID:
      name: project_id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/UUID"
...
    ListLimit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
    ListOffset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        maximum: 1000000
        default: 0
    ListSearch:
      name: q
      in: query
      schema:
        type: string
        maxLength: 253
  requestBodies:
    CreateNamedResource:
      required: true
      content:
        application/json:
          schema:
            type: object
            additionalProperties: false
            required: [name]
            properties:
              name:
                type: string
                minLength: 1
                maxLength: 253
  responses:
    InvalidRequest:
      description: 输入无效
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthenticated:
      description: 未认证或 Session 已失效
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: CSRF 或权限检查失败
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: 目标不存在或不可见
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: 幂等、生命周期或安全不变量冲突
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Error:
      type: object
      additionalProperties: false
      required: [code, message, request_id]
      properties:
        code:
          type: string
        message:
          type: string
        request_id:
          type: string
...
    Confirmation:
      type: object
      additionalProperties: false
      required: [confirm]
      properties:
        confirm:
          const: true
    UpdateScopedResourceRequest:
      type: object
      additionalProperties: false
      required: [name, status, confirm]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 253
        status:
          type: string
          enum: [active, suspended]
        confirm:
          type: boolean
          description: status=suspended 时必须为 true
...
    Pagination:
      type: object
      additionalProperties: false
      required: [limit, offset, total, has_more]
      properties:
        limit:
          type: integer
          minimum: 1
          maximum: 100
        offset:
          type: integer
          minimum: 0
        total:
          type: integer
          minimum: 0
        has_more:
          type: boolean
    Tenant:
      allOf:
        - $ref: "#/components/schemas/ResourceBase"
    Project:
      allOf:
        - $ref: "#/components/schemas/ResourceBase"
        - type: object
          required: [tenant_id]
          properties:
            tenant_id:
              $ref: "#/components/schemas/UUID"
...
    ClusterAggregate:
      allOf:
        - $ref: "#/components/schemas/Cluster"
        - type: object
          required: [connection]
          properties:
            connection:
              $ref: "#/components/schemas/ClusterConnection"
...